perf: move powersync sdk out of the entry chunk
