+++ package.json
@@ -178,6 +178,8 @@
     "oauth2-mock-server": "^8.2.2",
     "playwright": "^1.58.2",
     "prettier": "^3.6.2",
+    "rsocket-core": "1.0.0-alpha.3",
+    "rsocket-websocket-client": "1.0.0-alpha.3",
     "samlify": "^2.12.0",
     "size-limit": "^12.0.1",
     "storybook": "^10.3.5",
