+++ src/components/powersync-status.tsx
@@ -11,7 +11,7 @@ import { useSidebar } from '@/components/ui/sidebar'
 import { useSyncEnabledToggle } from '@/hooks/use-sync-enabled-toggle'
 import { edgeSpacing, mobileSidebarWidthRatio } from '@/lib/constants'
 import { cn } from '@/lib/utils'
-import { reconnectSync } from '@/db/powersync'
+import { reconnectSync } from '@/db/powersync/sync-state'
 import { Cloud, CloudOff, Loader2, RefreshCw } from 'lucide-react'
 import { SyncSetupModal } from '@/components/sync-setup/sync-setup-modal'
 import { Button } from '@/components/ui/button'
